+++ src/main/java/com/lgs/eden/api/nexus/AuthImp.java
@@ -4,6 +4,8 @@ import com.lgs.eden.api.APIException;
 import com.lgs.eden.api.APIResponseCode;
 import com.lgs.eden.api.auth.AuthAPI;
 import com.lgs.eden.api.auth.LoginResponseData;
+import com.lgs.eden.api.nexus.helpers.ImpSocket;
+import com.lgs.eden.api.nexus.helpers.MonitorIO;
 import io.socket.client.Ack;
 import io.socket.client.Socket;
 import org.json.JSONException;
