+++ src/main/java/com/lgs/eden/api/nexus/helpers/ImpSocket.java
@@ -1,4 +1,4 @@
-package com.lgs.eden.api.nexus;
+package com.lgs.eden.api.nexus.helpers;
 
 import com.lgs.eden.api.APIHandler;
 import io.socket.client.Socket;
@@ -13,4 +13,8 @@ public abstract class ImpSocket {
     }
 
     public void setParent(APIHandler parent) { this.parent = parent; }
+
+    public boolean notConnected() {
+        return !socket.connected();
+    }
 }
